pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["ble_exporter*"]
exclude = ["tests*", "logs*", "contrib*"]

[project]
name = "ble2"
version = "0.1.0"
description = "BLE Sensor Prometheus Exporter - passively listens for BLE advertisements and exposes sensor data to Prometheus"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bleak>=0.21.0",
    "cryptography>=47.0.0",
    "prometheus-client>=0.19.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    "coverage>=7.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--cov=ble_exporter",
    "--cov-report=html",
    "--cov-report=term",
    "--cov-report=term-missing",
    "-v",
]
filterwarnings = [
    "ignore::aiohttp.web_exceptions.NotAppKeyWarning",
]
